+++ src/components/Message/renderText/remarkPlugins/index.ts
@@ -1,2 +1,4 @@
 export * from './htmlToTextPlugin';
+export * from './imageToLink';
 export * from './keepLineBreaksPlugin';
+export * from './plusPlusToEmphasis';