+++ src/components/Message/renderText/renderText.tsx
@@ -10,7 +10,12 @@ import type { PluggableList } from 'unified'; // A sub-dependency of react-markd
 import { Anchor, Emoji, Mention } from './componentRenderers';
 import { detectHttp, matchMarkdownLinks, messageCodeBlocks } from './regex';
 import { emojiMarkdownPlugin, mentionsMarkdownPlugin } from './rehypePlugins';
-import { htmlToTextPlugin, keepLineBreaksPlugin } from './remarkPlugins';
+import {
+  htmlToTextPlugin,
+  imageToLink,
+  keepLineBreaksPlugin,
+  plusPlusToEmphasis,
+} from './remarkPlugins';
 import { ErrorBoundary } from '../../UtilityComponents';
 import type { MentionProps } from './componentRenderers';
 
@@ -45,6 +50,13 @@ export const defaultAllowedTagNames: Array<
   // custom types (tagNames)
   'emoji',
   'mention',
+  'h1',
+  'h2',
+  'h3',
+  'h4',
+  'h5',
+  'h6',
+  'ins',
 ];
 
 function formatUrlForDisplay(url: string) {
@@ -163,6 +175,8 @@ export const renderText = (
     htmlToTextPlugin,
     keepLineBreaksPlugin,
     [remarkGfm, { singleTilde: false }],
+    plusPlusToEmphasis,
+    imageToLink,
   ];
   const rehypePlugins: PluggableList = [emojiMarkdownPlugin];